Split connection actor tests and refine docs
